finished distance rendering

+++ backend/db/migrations/20201120213108-create-user.js
@@ -57,11 +57,11 @@ module.exports = {
         allowNull: false
       },
       lat: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.FLOAT(20),
         allowNull: false
       },
       lng: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.FLOAT(20),
         allowNull: false
       },
       createdAt: {
